+++ sdk/keyvault/azure-security-keyvault-keys/inc/azure/keyvault/keys/key_client_models.hpp
@@ -14,6 +14,7 @@
 #include <azure/core/context.hpp>
 #include <azure/core/datetime.hpp>
 #include <azure/core/http/http.hpp>
+#include <azure/core/internal/extendable_enumeration.hpp>
 #include <azure/core/nullable.hpp>
 #include <azure/core/operation.hpp>
 #include <azure/core/operation_status.hpp>
@@ -90,46 +91,52 @@ namespace Azure { namespace Security { namespace KeyVault { namespace Keys {
     std::string const& ToString() const { return m_operation; }
 
     /**
-     * @brief The key can be used to encrypt with the #Encrypt(EncryptionAlgorithm, Byte[],
-     * CancellationToken) method.
+     * @brief The key can be used to encrypt with the #Encrypt method.
+     *
      */
     AZ_SECURITY_KEYVAULT_KEYS_DLLEXPORT static const KeyOperation Encrypt;
 
     /**
-     * @brief The key can be used to decrypt with the #Decrypt(EncryptionAlgorithm, Byte[],
-     * CancellationToken) method.
+     * @brief The key can be used to decrypt with the #Decrypt method.
+     *
      */
     AZ_SECURITY_KEYVAULT_KEYS_DLLEXPORT static const KeyOperation Decrypt;
 
     /**
-     * @brief The key can be used to sign with the Sign(SignatureAlgorithm, Byte[],
-     * CancellationToken) method.
+     * @brief The key can be used to sign with the Sign method.
+     *
      */
     AZ_SECURITY_KEYVAULT_KEYS_DLLEXPORT static const KeyOperation Sign;
 
     /**
-     * @brief The key can be used to verify with the Verify(SignatureAlgorithm, Byte[], Byte[],
-     * CancellationToken) method.
+     * @brief The key can be used to verify with the Verify method.
+     *
      */
     AZ_SECURITY_KEYVAULT_KEYS_DLLEXPORT static const KeyOperation Verify;
 
     /**
-     * @brief The key can be used to wrap another key with the WrapKey(KeyWrapAlgorithm, Byte[],
-     * CancellationToken) method.
+     * @brief The key can be used to wrap another key with the WrapKey method.
+     *
      */
     AZ_SECURITY_KEYVAULT_KEYS_DLLEXPORT static const KeyOperation WrapKey;
 
     /**
-     * @brief The key can be used to unwrap another key with the UnwrapKey(KeyWrapAlgorithm, Byte[],
-     * CancellationToken) method.
+     * @brief The key can be used to unwrap another key with the UnwrapKey method.
+     *
      */
     AZ_SECURITY_KEYVAULT_KEYS_DLLEXPORT static const KeyOperation UnwrapKey;
 
     /**
-     * @brief The key can be imported during creation using the ImportKey(ImportKeyOptions,
-     * CancellationToken) method.
+     * @brief The key can be imported during creation using the ImportKey method.
+     *
      */
     AZ_SECURITY_KEYVAULT_KEYS_DLLEXPORT static const KeyOperation Import;
+
+    /**
+     * @brief The key can be exported during creation using the ImportKey method.
+     *
+     */
+    AZ_SECURITY_KEYVAULT_KEYS_DLLEXPORT static const KeyOperation Export;
   };
 
   /**
@@ -293,6 +300,39 @@ namespace Azure { namespace Security { namespace KeyVault { namespace Keys {
   };
 
   /**
+   * @brief The encryption algorithm to use to protect the exported key material.
+   *
+   */
+  class KeyEncryptionAlgorithm final
+      : public Azure::Core::_internal::ExtendableEnumeration<KeyEncryptionAlgorithm> {
+  public:
+    /**
+     * @brief Construct a new KeyEncryptionAlgorithm object.
+     *
+     * @param value The string value of the instance.
+     */
+    explicit KeyEncryptionAlgorithm(std::string value) : ExtendableEnumeration(std::move(value)) {}
+
+    /**
+     * @brief Gets the CKM_RSA_AES_KEY_WRAP algorithm.
+     *
+     */
+    AZ_SECURITY_KEYVAULT_KEYS_DLLEXPORT static const KeyEncryptionAlgorithm CKM_RSA_AES_KEY_WRAP;
+
+    /**
+     * @brief Gets the RSA_AES_KEY_WRAP_256 algorithm.
+     *
+     */
+    AZ_SECURITY_KEYVAULT_KEYS_DLLEXPORT static const KeyEncryptionAlgorithm RSA_AES_KEY_WRAP_256;
+
+    /**
+     * @brief Gets the RSA_AES_KEY_WRAP_384 algorithm.
+     *
+     */
+    AZ_SECURITY_KEYVAULT_KEYS_DLLEXPORT static const KeyEncryptionAlgorithm RSA_AES_KEY_WRAP_384;
+  };
+
+  /**
    * @brief Represents a JSON Web Key as defined in http://tools.ietf.org/html/rfc7517.
    *
    */
@@ -389,6 +429,34 @@ namespace Azure { namespace Security { namespace KeyVault { namespace Keys {
   };
 
   /**
+   * @brief Key Release Policy
+   *
+   */
+  struct KeyReleasePolicy final
+  {
+    /**
+     * @brief Content type and version of key release policy.
+     *
+     * @remark Default value: "application/json; charset=utf-8"
+     */
+    Azure::Nullable<std::string> ContentType;
+
+    /**
+     * @brief Defines the mutability state of the policy. Once marked immutable, this flag cannot be
+     * reset and the policy cannot be changed under any circumstances.
+     *
+     */
+    bool Immutable;
+
+    /**
+     * @brief Blob encoding the policy rules under which the key can be released.
+     *
+     * @remark Format: base64url
+     */
+    std::string EncodedPolicy;
+  };
+
+  /**
    * @brief The resource containing all the properties of the KeyVaultKey except JsonWebKey
    * properties.
    *
@@ -420,8 +488,8 @@ namespace Azure { namespace Security { namespace KeyVault { namespace Keys {
     std::string Version;
 
     /**
-     * @brief Indicate whether the key's lifetime is managed by Key Vault. If this key is backing a
-     * Key Vault certificate, the value will be true.
+     * @brief Indicate whether the key's lifetime is managed by Key Vault. If this key is backing
+     * a Key Vault certificate, the value will be true.
      *
      */
     bool Managed = false;
@@ -479,6 +547,18 @@ namespace Azure { namespace Security { namespace KeyVault { namespace Keys {
     std::string RecoveryLevel;
 
     /**
+     * @brief The policy rules under which the key can be exported.
+     *
+     */
+    Azure::Nullable<KeyReleasePolicy> ReleasePolicy;
+
+    /**
+     * @brief Indicates if the private key can be exported.
+     *
+     */
+    Azure::Nullable<bool> Exportable;
+
+    /**
      * @brief Construct a new Key Properties object.
      *
      */
@@ -505,7 +585,8 @@ namespace Azure { namespace Security { namespace KeyVault { namespace Keys {
     virtual ~KeyVaultKey() = default;
 
     /**
-     * @brief The cryptographic key, the key type, and the operations you can perform using the key.
+     * @brief The cryptographic key, the key type, and the operations you can perform using the
+     * key.
      *
      */
     JsonWebKey Key;
@@ -670,7 +751,8 @@ namespace Azure { namespace Security { namespace KeyVault { namespace Keys {
      *
      * @param deletedKeyProperties A previously created #DeletedKeyPagedResponse that is used to
      * init this new instance.
-     * @param rawResponse The HTTP raw response from where the #DeletedKeyPagedResponse was parsed.
+     * @param rawResponse The HTTP raw response from where the #DeletedKeyPagedResponse was
+     * parsed.
      * @param keyClient A key client required for getting the next pages.
      */
     DeletedKeyPagedResponse(
@@ -713,8 +795,8 @@ namespace Azure { namespace Security { namespace KeyVault { namespace Keys {
     std::string m_continuationToken;
 
     /* This is the implementation for checking the status of a deleted key. The key is considered
-     * deleted if querying /deletedkeys/keyName returns 200 from server. Or whenever soft-delete is
-     * disabled.*/
+     * deleted if querying /deletedkeys/keyName returns 200 from server. Or whenever soft-delete
+     * is disabled.*/
     std::unique_ptr<Azure::Core::Http::RawResponse> PollInternal(
         Azure::Core::Context const& context) override;
 
@@ -776,7 +858,8 @@ namespace Azure { namespace Security { namespace KeyVault { namespace Keys {
     Azure::Security::KeyVault::Keys::DeletedKey Value() const override { return m_value; }
 
     /**
-     * @brief Get an Url as string which can be used to get the status of the delete key operation.
+     * @brief Get an Url as string which can be used to get the status of the delete key
+     * operation.
      *
      * @return std::string
      */
@@ -789,7 +872,8 @@ namespace Azure { namespace Security { namespace KeyVault { namespace Keys {
      * @remark After the operation is initialized, it is used to poll the last update from the
      * server using the \p context.
      *
-     * @param resumeToken A previously generated token used to resume the polling of the operation.
+     * @param resumeToken A previously generated token used to resume the polling of the
+     * operation.
      * @param client A #KeyClient that is used for getting status updates.
      * @param context A #Azure::Core::Context controlling the request lifetime.
      * @return DeleteKeyOperation
@@ -837,8 +921,8 @@ namespace Azure { namespace Security { namespace KeyVault { namespace Keys {
     }
 
     /*
-     * Only friend classes are permitted to construct a RecoverDeletedKeyOperation. This is because
-     * a KeyVaultPipelne is required and it is not exposed to customers.
+     * Only friend classes are permitted to construct a RecoverDeletedKeyOperation. This is
+     * because a KeyVaultPipelne is required and it is not exposed to customers.
      *
      * Since C++ doesn't offer `internal` access, we use friends-only instead.
      */
@@ -875,7 +959,8 @@ namespace Azure { namespace Security { namespace KeyVault { namespace Keys {
     Azure::Security::KeyVault::Keys::KeyVaultKey Value() const override { return m_value; }
 
     /**
-     * @brief Get an Url as string which can be used to get the status of the delete key operation.
+     * @brief Get an Url as string which can be used to get the status of the delete key
+     * operation.
      *
      * @return std::string
      */
@@ -888,7 +973,8 @@ namespace Azure { namespace Security { namespace KeyVault { namespace Keys {
      * @remark After the operation is initialized, it is used to poll the last update from the
      * server using the \p context.
      *
-     * @param resumeToken A previously generated token used to resume the polling of the operation.
+     * @param resumeToken A previously generated token used to resume the polling of the
+     * operation.
      * @param client A #KeyClient that is used for getting status updates.
      * @param context A #Azure::Core::Context controlling the request lifetime.
      * @return DeleteKeyOperation
@@ -899,4 +985,136 @@ namespace Azure { namespace Security { namespace KeyVault { namespace Keys {
         Azure::Core::Context const& context = Azure::Core::Context());
   };
 
+  /**
+   * @brief LifeTime action type
+   *
+   */
+  enum class LifetimeActionType
+  {
+    /**
+     * @brief Rotate the key based on the key policy.
+     *
+     */
+    Rotate,
+
+    /**
+     * @brief Trigger event grid events. For preview, the notification time is not configurable
+     * and it is default to 30 days before expiry.
+     *
+     */
+    Notify
+  };
+
+  /**
+   * @brief A condition to be satisfied for an action to be executed.
+   */
+  struct LifetimeActionsTrigger final
+  {
+    /**
+     * @brief Time after creation to attempt to rotate. It only applies to rotate. It will be in
+     * ISO 8601 duration format. Example: 90 days : "P90D"
+     *
+     */
+    Azure::Nullable<std::string> TimeAfterCreate;
+    /**
+     * @brief Time before expiry to attempt to rotate or notify. It will be in ISO 8601 duration
+     * format. Example: 90 days : "P90D"
+     *
+     */
+    Azure::Nullable<std::string> TimeBeforeExpiry;
+  };
+
+  /**
+   * @brief Action and its trigger that will be performed by Key Vault over the lifetime of a key.
+   *
+   */
+  struct LifetimeActionsType final
+  {
+    /**
+     * @brief The condition that will execute the action.
+     *
+     */
+    LifetimeActionsTrigger Trigger;
+
+    /**
+     * @brief The action that will be executed.
+     */
+    LifetimeActionType Action;
+  };
+
+  /**
+   * @brief The key rotation policy attributes.
+   *
+   */
+  struct KeyRotationPolicyAttributes final
+  {
+    /**
+     * @brief The expiryTime will be applied on the new key version. It should be at least 28
+     * days. It will be in ISO 8601 Format. Examples: 90 days: P90D, 3 months: P3M, 48 hours:
+     * PT48H, 1 year and 10 days: P1Y10D
+     */
+    Azure::Nullable<std::string> ExpiryTime;
+
+    /**
+     * @brief The key rotation policy created time in UTC.
+     *
+     */
+    Azure::Nullable<Azure::DateTime> Created;
+
+    /**
+     * @brief The key rotation policy's last updated time in UTC.
+     *
+     */
+    Azure::Nullable<Azure::DateTime> Updated;
+  };
+
+  /**
+   * @brief Rotation policy for a key.
+   */
+  struct KeyRotationPolicy final
+  {
+    /**
+     * @brief The key policy id.
+     */
+    std::string Id;
+
+    /**
+     * @brief Actions that will be performed by Key Vault over the lifetime of a key. For preview,
+     * lifetimeActions can only have two items at maximum: one for rotate, one for notify.
+     * Notification time would be default to 30 days before expiry and it is not configurable.
+     *
+     */
+    std::vector<LifetimeActionsType> LifetimeActions;
+
+    /**
+     * @brief The key rotation policy attributes.
+     */
+    KeyRotationPolicyAttributes Attributes;
+  };
+
+  /**
+   * @brief The GetRandomBytes result type containing the random bytes bytes.
+   *
+   */
+  struct GetRandomBytesResult final
+  {
+    /**
+     * @brief The random generated bytes.
+     *
+     */
+    std::vector<uint8_t> RandomBytes;
+  };
+
+  /**
+   * @brief The release result, containing the released key.
+   *
+   */
+  struct ReleaseKeyResult
+  {
+    /**
+     * @brief A signed object containing the released key.
+     *
+     */
+    std::string Value;
+  };
 }}}} // namespace Azure::Security::KeyVault::Keys