+++ sdk/keyvault/azure-security-keyvault-certificates/CHANGELOG.md
@@ -4,11 +4,11 @@
 
 ### Features Added
 
+- Keyvault 7.3 support added for Certificates.
+
 ### Breaking Changes
 
-### Bugs Fixed
-
-### Other Changes
+- Removed ServiceVersion type, replaced with Version field in the CertificateClientOptions type.
 
 ## 4.0.0 (2022-06-07)